renamed db to reflect this is specific to postgresQL This will allow us to use different DBs (mongoDB) in the future
