+++ src/server/server.js
@@ -1,10 +1,10 @@
 const { appCreator } = require('./appCreator.js');
 const PORT = 8080;
 
-const { db } = require('../db/db.js');
+const dbPostgreSQL = require('../db/dbPostgreSQL.js');
 
 // start server
-const app = appCreator(db);
+const app = appCreator(dbPostgreSQL);
 
 app.listen(PORT, () => {
   console.log(`Server listening on port: ${PORT}...`);
